Moved tasks into a separate file
