+++ exotica/src/Problems/UnconstrainedEndPoseProblem.cpp
@@ -40,74 +40,6 @@ REGISTER_PROBLEM_TYPE("UnconstrainedEndPoseProblem", exotica::UnconstrainedEndPo
 namespace exotica
 {
 
-EndPoseTask::EndPoseTask()
-{
-
-}
-
-void EndPoseTask::initialize(const std::vector<exotica::Initializer>& inits, PlanningProblem_ptr prob, TaskSpaceVector& phi)
-{
-    Task::initialize(inits, prob, Phi);
-    y=Phi;
-    y.setZero(PhiN);
-    Rho = Eigen::VectorXd::Ones(NumTasks);
-    J = Eigen::MatrixXd(JN, prob->N);
-    S = Eigen::MatrixXd::Identity(JN, JN);
-    ydiff = Eigen::VectorXd::Zero(JN);
-
-    for(int i=0; i<NumTasks; i++)
-    {
-        TaskInitializer task(inits[i]);
-        if(task.Goal.rows()==0)
-        {
-            // Keep zero goal
-        }
-        else if(task.Goal.rows()==Tasks[i]->Length)
-        {
-            y.data.segment(Tasks[i]->Start, Tasks[i]->Length) = task.Goal;
-        }
-        else
-        {
-            throw_pretty("Invalid task goal size! Expecting "<<Tasks[i]->Length<<" got "<<task.Goal.rows());
-        }
-        if(task.Rho.rows()==0)
-        {
-            Rho(i) = 1.0;
-        }
-        else if(task.Rho.rows()==1)
-        {
-            Rho(i) = task.Rho(0);
-        }
-        else
-        {
-            throw_pretty("Invalid task Rho size! Expecting 1 got "<<task.Rho.rows());
-        }
-    }
-}
-
-void EndPoseTask::updateS()
-{
-    for (const TaskIndexing& task : Indexing)
-    {
-        for (int i = 0; i < task.Length; i++)
-        {
-            S(i + task.Start, i + task.Start) = Rho(task.Id);
-            if(Rho(task.Id)>0.0) Tasks[task.Id]->isUsed = true;
-        }
-    }
-}
-
-void EndPoseTask::update(const TaskSpaceVector& bigPhi, Eigen::MatrixXdRefConst bigJ)
-{
-    for (const TaskIndexing& task : Indexing)
-    {
-        Phi.data.segment(task.Start, task.Length) = bigPhi.data.segment(Tasks[task.Id]->Start, Tasks[task.Id]->Length);
-        J.middleRows(task.StartJ, task.LengthJ) = bigJ.middleRows(Tasks[task.Id]->StartJ, Tasks[task.Id]->LengthJ);
-    }
-    ydiff = Phi - y;
-}
-
-
 UnconstrainedEndPoseProblem::UnconstrainedEndPoseProblem()
 {
     Flags = KIN_FK | KIN_J;
