+++ exotica/src/Problems/UnconstrainedTimeIndexedProblem.cpp
@@ -39,84 +39,6 @@ REGISTER_PROBLEM_TYPE("UnconstrainedTimeIndexedProblem", exotica::UnconstrainedT
 namespace exotica
 {
 
-TimeIndexedTask::TimeIndexedTask()
-{
-
-}
-
-void TimeIndexedTask::initialize(const std::vector<exotica::Initializer>& inits, PlanningProblem_ptr prob, TaskSpaceVector& phi)
-{
-    Task::initialize(inits, prob, phi);
-    T=std::static_pointer_cast<UnconstrainedTimeIndexedProblem>(prob)->T;
-    phi.setZero(PhiN);
-    Phi.assign(T, phi);
-    y=Phi;
-    Rho.assign(T, Eigen::VectorXd::Ones(NumTasks));
-    J.assign(T, Eigen::MatrixXd(JN, prob->N));
-    S.assign(T, Eigen::MatrixXd::Identity(JN, JN));
-    ydiff.assign(T, Eigen::VectorXd::Zero(JN));
-
-    for(int i=0; i<NumTasks; i++)
-    {
-        TaskInitializer task(inits[i]);
-        if(task.Goal.rows()==0)
-        {
-            // Keep zero goal
-        }
-        else if(task.Goal.rows()==Tasks[i]->Length*T)
-        {
-            for(int t=0; t<T; t++)
-            {
-                y[t].data.segment(Tasks[i]->Start, Tasks[i]->Length) = task.Goal.segment(t*Tasks[i]->Length, Tasks[i]->Length);
-            }
-        }
-        else
-        {
-            throw_pretty("Invalid task goal size! Expecting "<<Tasks[i]->Length<<" got "<<task.Goal.rows());
-        }
-        if(task.Rho.rows()==0)
-        {
-            // Keeo ones
-        }
-        else if(task.Rho.rows()==T)
-        {
-            for(int t=0; t<T; t++)
-            {
-                Rho[t](i) = task.Rho(t);
-            }
-        }
-        else
-        {
-            throw_pretty("Invalid task Rho size! Expecting 1 got "<<task.Rho.rows());
-        }
-    }
-}
-
-void TimeIndexedTask::updateS()
-{
-    for(int t=0; t<T; t++)
-    {
-        for (const TaskIndexing& task : Indexing)
-        {
-            for (int i = 0; i < task.Length; i++)
-            {
-                S[t](i + task.Start, i + task.Start) = Rho[t](task.Id);
-                if(Rho[t](task.Id)>0.0) Tasks[task.Id]->isUsed = true;
-            }
-        }
-    }
-}
-
-void TimeIndexedTask::update(const TaskSpaceVector& bigPhi, Eigen::MatrixXdRefConst bigJ, int t)
-{
-    for (const TaskIndexing& task : Indexing)
-    {
-        Phi[t].data.segment(task.Start, task.Length) = bigPhi.data.segment(Tasks[task.Id]->Start, Tasks[task.Id]->Length);
-        J[t].middleRows(task.StartJ, task.LengthJ) = bigJ.middleRows(Tasks[task.Id]->StartJ, Tasks[task.Id]->LengthJ);
-    }
-    ydiff[t] = Phi[t] - y[t];
-}
-
 UnconstrainedTimeIndexedProblem::UnconstrainedTimeIndexedProblem()
     : T(0), tau(0), Q_rate(0), W_rate(0), H_rate(0)
 {
